Split agent session discovery
