+++ rust/src/agent_sessions.rs
@@ -294,539 +294,9 @@ pub struct AgentSessionDiscovery {
     remote: RemoteSessionFetcher,
 }
 
-impl AgentPSOutputParser {
-    pub fn parse(output: &str) -> Vec<AgentProcessRecord> {
-        let mut seen_pids = HashSet::new();
-        output
-            .lines()
-            .filter_map(|line| Self::parse_line(line, &mut seen_pids))
-            .collect()
-    }
-
-    pub fn agent_processes(records: &[AgentProcessRecord]) -> Vec<AgentProcessRecord> {
-        let mut seen = HashSet::new();
-        records
-            .iter()
-            .filter(|record| record.is_agent())
-            .filter_map(|record| {
-                if seen.insert(record.pid) {
-                    Some(record.clone())
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    pub fn provider(record: &AgentProcessRecord) -> Option<AgentSessionProvider> {
-        record.provider
-    }
-
-    pub fn source(record: &AgentProcessRecord) -> AgentSessionSource {
-        record.source
-    }
-
-    pub fn has_codex_app_server(records: &[AgentProcessRecord]) -> bool {
-        records.iter().any(|record| {
-            record.kind == AgentProcessKind::AppServer
-                && record.provider == Some(AgentSessionProvider::Codex)
-        })
-    }
-
-    fn parse_line(line: &str, seen_pids: &mut HashSet<u32>) -> Option<AgentProcessRecord> {
-        let mut fields = line.split_whitespace();
-        let pid = fields.next()?.parse::<u32>().ok()?;
-        let ppid = fields.next()?.parse::<u32>().ok()?;
-        let weekday = fields.next()?;
-        let month = fields.next()?;
-        let day = fields.next()?;
-        let time = fields.next()?;
-        let year = fields.next()?;
-        if !seen_pids.insert(pid) {
-            return None;
-        }
-
-        let started_at = Self::parse_started_at(weekday, month, day, time, year)?;
-        let command = fields.collect::<Vec<_>>().join(" ");
-        let classification = classify_process_command(&command);
-        Some(AgentProcessRecord {
-            pid,
-            ppid,
-            started_at: Some(started_at),
-            provider: classification.provider,
-            source: classification.source,
-            executable: classification.executable,
-            kind: classification.kind,
-        })
-    }
-
-    fn parse_started_at(
-        weekday: &str,
-        month: &str,
-        day: &str,
-        time: &str,
-        year: &str,
-    ) -> Option<DateTime<Utc>> {
-        let text = format!("{weekday} {month} {day} {time} {year}");
-        chrono::NaiveDateTime::parse_from_str(&text, "%a %b %e %H:%M:%S %Y")
-            .ok()
-            .map(|dt| DateTime::<Utc>::from_naive_utc_and_offset(dt, Utc))
-    }
-}
-
-#[derive(Deserialize)]
-#[serde(rename_all = "PascalCase")]
-struct WindowsProcessMetadata {
-    process_id: u32,
-    #[serde(default)]
-    parent_process_id: u32,
-    creation_date: Option<String>,
-    name: Option<String>,
-    executable_path: Option<String>,
-}
-
-impl WindowsProcessOutputParser {
-    pub fn parse(output: &str) -> Vec<AgentProcessRecord> {
-        let Ok(value) = serde_json::from_str::<Value>(output.trim()) else {
-            return Vec::new();
-        };
-        let values = match value {
-            Value::Array(values) => values,
-            Value::Object(_) => vec![value],
-            _ => return Vec::new(),
-        };
-        let mut seen = HashSet::new();
-
-        values
-            .into_iter()
-            .filter_map(|value| serde_json::from_value::<WindowsProcessMetadata>(value).ok())
-            .filter(|process| process.process_id > 0 && seen.insert(process.process_id))
-            .map(|process| {
-                let display_name = process
-                    .name
-                    .as_deref()
-                    .filter(|name| !name.trim().is_empty())
-                    .or(process.executable_path.as_deref())
-                    .unwrap_or_default();
-                let classification = classify_process_command(display_name);
-                AgentProcessRecord {
-                    pid: process.process_id,
-                    ppid: process.parent_process_id,
-                    started_at: process
-                        .creation_date
-                        .as_deref()
-                        .and_then(parse_windows_creation_date),
-                    provider: classification.provider,
-                    source: classification.source,
-                    executable: process.name.unwrap_or(classification.executable),
-                    kind: classification.kind,
-                }
-            })
-            .collect()
-    }
-}
-
-fn parse_windows_creation_date(value: &str) -> Option<DateTime<Utc>> {
-    DateTime::parse_from_rfc3339(value)
-        .ok()
-        .map(|date| date.with_timezone(&Utc))
-        .or_else(|| {
-            let value = value.strip_prefix("/Date(")?;
-            let milliseconds = value
-                .trim_end_matches(")/")
-                .split(['+', '-'])
-                .next()?
-                .parse()
-                .ok()?;
-            DateTime::<Utc>::from_timestamp_millis(milliseconds)
-        })
-        .or_else(|| {
-            let core = value.get(..21)?;
-            chrono::NaiveDateTime::parse_from_str(core, "%Y%m%d%H%M%S%.f")
-                .ok()
-                .map(|date| DateTime::<Utc>::from_naive_utc_and_offset(date, Utc))
-        })
-}
-
-struct ProcessClassification {
-    provider: Option<AgentSessionProvider>,
-    source: AgentSessionSource,
-    executable: String,
-    kind: AgentProcessKind,
-}
-
-fn classify_process_command(command: &str) -> ProcessClassification {
-    let lower = command.to_ascii_lowercase();
-    let executable = executable_basename(command);
-
-    if lower.contains("app-server") && lower.contains("codex") {
-        return ProcessClassification {
-            provider: Some(AgentSessionProvider::Codex),
-            source: AgentSessionSource::DesktopApp,
-            executable,
-            kind: AgentProcessKind::AppServer,
-        };
-    }
-
-    if lower.contains("codex (renderer)")
-        || lower.contains("claude-code-acp")
-        || lower.contains("--help")
-        || lower.contains("--version")
-        || lower.contains("--type=renderer")
-        || lower.contains("disclaimer")
-        || executable.eq_ignore_ascii_case("disclaimer")
-    {
-        return ProcessClassification {
-            provider: None,
-            source: AgentSessionSource::Unknown,
-            executable,
-            kind: AgentProcessKind::Helper,
-        };
-    }
-
-    if lower.contains("application support/claude/claude-code/claude")
-        || lower.contains("claude.app")
-        || lower.contains("claude.exe")
-        || executable.eq_ignore_ascii_case("claude")
-    {
-        return ProcessClassification {
-            provider: Some(AgentSessionProvider::Claude),
-            source: if lower.contains("application support/claude/claude-code")
-                || lower.contains("claude.app")
-            {
-                AgentSessionSource::DesktopApp
-            } else {
-                AgentSessionSource::Cli
-            },
-            executable: if executable.eq_ignore_ascii_case("claude") {
-                "claude".to_string()
-            } else {
-                executable
-            },
-            kind: AgentProcessKind::Agent,
-        };
-    }
-
-    if lower.contains("codex.exe")
-        || lower.contains("codex.app")
-        || lower.contains("codex desktop")
-        || executable.eq_ignore_ascii_case("codex")
-    {
-        return ProcessClassification {
-            provider: Some(AgentSessionProvider::Codex),
-            source: if lower.contains("codex.app") || lower.contains("codex desktop") {
-                AgentSessionSource::DesktopApp
-            } else {
-                AgentSessionSource::Cli
-            },
-            executable: if executable.eq_ignore_ascii_case("codex") {
-                "codex".to_string()
-            } else {
-                executable
-            },
-            kind: AgentProcessKind::Agent,
-        };
-    }
-
-    ProcessClassification {
-        provider: None,
-        source: AgentSessionSource::Unknown,
-        executable,
-        kind: AgentProcessKind::Other,
-    }
-}
-
-fn executable_basename(command: &str) -> String {
-    let normalized = command.replace('\\', "/").to_ascii_lowercase();
-    for (needle, name) in [
-        ("claude-code-acp", "claude-code-acp"),
-        ("application support/claude/claude-code/claude", "claude"),
-        ("codex app-server", "codex"),
-        ("codex (renderer)", "codex"),
-        ("codex.app", "codex"),
-        ("claude.app", "claude"),
-        ("claude.exe", "claude"),
-        ("codex.exe", "codex"),
-        ("disclaimer", "disclaimer"),
-    ] {
-        if normalized.contains(needle) {
-            return name.to_string();
-        }
-    }
-
-    let first_token = command.split_whitespace().next().unwrap_or_default();
-    if first_token.is_empty() {
-        return String::new();
-    }
-
-    Path::new(first_token)
-        .file_name()
-        .and_then(|name| name.to_str())
-        .unwrap_or(first_token)
-        .to_string()
-}
-
-impl LSOFCWDOutputParser {
-    pub fn parse(output: &str) -> HashMap<u32, String> {
-        let mut result = HashMap::new();
-        let mut current_pid = None;
-
-        for line in output.lines() {
-            let line = line.trim();
-            if line.is_empty() {
-                continue;
-            }
-
-            match line.chars().next() {
-                Some('p') => {
-                    current_pid = line[1..].trim().parse::<u32>().ok();
-                }
-                Some('n') => {
-                    if let Some(pid) = current_pid {
-                        result.insert(pid, line[1..].to_string());
-                    }
-                }
-                _ => {}
-            }
-        }
-
-        result
-    }
-}
-
-impl ClaudeSessionProjectMapper {
-    pub fn escaped_cwd(cwd: &str) -> String {
-        cwd.chars()
-            .map(|scalar| {
-                if scalar.is_ascii_alphanumeric() {
-                    scalar
-                } else {
-                    '-'
-                }
-            })
-            .collect()
-    }
-
-    pub fn project_directories(cwd: &str, home_directory: &Path) -> Vec<PathBuf> {
-        if cwd.trim().is_empty() {
-            return Vec::new();
-        }
-
-        vec![
-            home_directory
-                .join(".claude")
-                .join("projects")
-                .join(Self::escaped_cwd(cwd)),
-        ]
-    }
-
-    pub fn transcripts(cwd: &str, home_directory: &Path) -> Vec<ClaudeTranscript> {
-        let mut transcripts = Vec::new();
-
-        for directory in Self::project_directories(cwd, home_directory) {
-            let Ok(entries) = fs::read_dir(&directory) else {
-                continue;
-            };
-
-            for entry in entries.flatten() {
-                let path = entry.path();
-                if path.extension().and_then(|ext| ext.to_str()) != Some("jsonl") {
-                    continue;
-                }
-
-                let Ok(metadata) = entry.metadata() else {
-                    continue;
-                };
-                let Ok(modified) = metadata.modified() else {
-                    continue;
-                };
-
-                transcripts.push(ClaudeTranscript::new(path, modified.into()));
-            }
-        }
-
-        transcripts.sort_by(|lhs, rhs| {
-            rhs.modified_at
-                .cmp(&lhs.modified_at)
-                .then_with(|| rhs.url.cmp(&lhs.url))
-        });
-        transcripts
-    }
-
-    pub fn newest_transcript(cwd: &str, home_directory: &Path) -> Option<ClaudeTranscript> {
-        Self::transcripts(cwd, home_directory).into_iter().next()
-    }
-}
-
-impl ClaudeTranscriptMetadataParser {
-    const MAX_LINES: usize = 32;
-    const MAX_BYTES: u64 = 64 * 1024;
-
-    pub fn parse(reader: impl Read) -> Option<ClaudeTranscriptMetadata> {
-        let mut session_id = None;
-        let mut cwd = None;
-        let reader = BufReader::new(reader.take(Self::MAX_BYTES));
-
-        for line in reader.lines().take(Self::MAX_LINES).map_while(Result::ok) {
-            let Ok(value) = serde_json::from_str::<Value>(&line) else {
-                continue;
-            };
-            if session_id.is_none() {
-                session_id = value
-                    .get("sessionId")
-                    .or_else(|| value.get("session_id"))
-                    .and_then(Value::as_str)
-                    .map(str::to_owned);
-            }
-            if cwd.is_none() {
-                cwd = value.get("cwd").and_then(Value::as_str).map(str::to_owned);
-            }
-            if session_id.is_some() && cwd.is_some() {
-                break;
-            }
-        }
-
-        (session_id.is_some() || cwd.is_some())
-            .then_some(ClaudeTranscriptMetadata { session_id, cwd })
-    }
-}
-
-impl CodexRolloutFirstLineParser {
-    pub fn parse(line: &str) -> Option<CodexRolloutMetadata> {
-        let value: Value = serde_json::from_str(line).ok()?;
-        if value.get("type")?.as_str()? != "session_meta" {
-            return None;
-        }
-
-        let payload = value.get("payload")?.as_object()?;
-        let session_id = payload
-            .get("session_id")
-            .or_else(|| payload.get("id"))?
-            .as_str()?;
-        let session_id = session_id.trim();
-        if session_id.is_empty() {
-            return None;
-        }
-
-        Some(CodexRolloutMetadata {
-            session_id: session_id.to_string(),
-            cwd: payload
-                .get("cwd")
-                .and_then(|value| value.as_str())
-                .map(ToOwned::to_owned),
-            originator: payload
-                .get("originator")
-                .and_then(|value| value.as_str())
-                .map(ToOwned::to_owned),
-            source: payload
-                .get("source")
-                .and_then(|value| value.as_str())
-                .map(ToOwned::to_owned),
-        })
-    }
-
-    pub fn read_first_line(path: &Path) -> Option<String> {
-        let file = File::open(path).ok()?;
-        let mut reader = BufReader::new(file);
-        let mut line = String::new();
-        let bytes = reader.read_line(&mut line).ok()?;
-        if bytes == 0 {
-            return None;
-        }
-        while line.ends_with(['\n', '\r']) {
-            line.pop();
-        }
-        Some(line)
-    }
-}
-
-impl AgentSessionCorrelation {
-    pub fn project_name(cwd: Option<&str>) -> Option<String> {
-        cwd.and_then(project_name_from_cwd)
-    }
-}
-
-pub fn focus_session(session: &AgentSession) -> SessionFocusResult {
-    match session.focus_target {
-        AgentSessionFocusTarget::Transcript { .. } => SessionFocusResult::unsupported(
-            "This file-only session has no focusable Windows window.",
-        ),
-        AgentSessionFocusTarget::None => {
-            SessionFocusResult::unsupported("This session has no focus target on Windows.")
-        }
-        AgentSessionFocusTarget::Process { pid } => {
-            if !is_local_host(&session.host) {
-                return SessionFocusResult::unsupported(
-                    "Remote session focus is not supported from this Windows desktop.",
-                );
-            }
-
-            focus_process(pid)
-        }
-    }
-}
-
-fn is_local_host(host: &str) -> bool {
-    if host.eq_ignore_ascii_case("localhost")
-        || host == "127.0.0.1"
-        || host == "::1"
-        || host.is_empty()
-    {
-        return true;
-    }
-
-    std::env::var("COMPUTERNAME")
-        .map(|name| name.eq_ignore_ascii_case(host))
-        .unwrap_or(false)
-}
-
-#[cfg(windows)]
-fn focus_process(pid: u32) -> SessionFocusResult {
-    use windows::Win32::Foundation::{BOOL, HWND, LPARAM};
-    use windows::Win32::UI::WindowsAndMessaging::{
-        EnumWindows, GetWindowThreadProcessId, IsWindowVisible, SW_RESTORE, SetForegroundWindow,
-        ShowWindow,
-    };
-
-    struct Search {
-        pid: u32,
-        window: Option<HWND>,
-    }
-
-    unsafe extern "system" fn find_window(hwnd: HWND, data: LPARAM) -> BOOL {
-        let search = unsafe { &mut *(data.0 as *mut Search) };
-        let mut window_pid = 0;
-        unsafe { GetWindowThreadProcessId(hwnd, Some(&mut window_pid)) };
-        if window_pid == search.pid && unsafe { IsWindowVisible(hwnd).as_bool() } {
-            search.window = Some(hwnd);
-            return BOOL(0);
-        }
-        BOOL(1)
-    }
-
-    let mut search = Search { pid, window: None };
-    let result = unsafe { EnumWindows(Some(find_window), LPARAM(&mut search as *mut _ as isize)) };
-    if result.is_err() {
-        return SessionFocusResult::failed("Windows could not enumerate application windows.");
-    }
-
-    let Some(window) = search.window else {
-        return SessionFocusResult::failed("No focusable window was found for this session.");
-    };
-    unsafe {
-        let _ = ShowWindow(window, SW_RESTORE);
-        if SetForegroundWindow(window).as_bool() {
-            SessionFocusResult::focused()
-        } else {
-            SessionFocusResult::failed("Windows denied the request to focus this session.")
-        }
-    }
-}
-
-#[cfg(not(windows))]
-fn focus_process(_pid: u32) -> SessionFocusResult {
-    SessionFocusResult::unsupported("Process focus requires the Windows desktop shell.")
-}
+mod focus;
+mod parsers;
+pub use focus::focus_session;
 
 struct CodexRollout {
     path: PathBuf,
@@ -1476,352 +946,4 @@ fn project_name_from_cwd(cwd: &str) -> Option<String> {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use chrono::TimeZone;
-    use std::io;
-    use std::sync::Arc;
-
-    #[test]
-    fn process_parser_filters_helpers_app_server_duplicates_and_malformed_lines() {
-        let output = "\
-101   1 Mon Jul  6 09:00:00 2026 /Applications/Claude.app/Contents/Resources/disclaimer /Users/test/Library/Application Support/Claude/claude-code/claude --dangerously-skip-permissions
-102 101 Mon Jul  6 09:00:01 2026 /Users/test/Library/Application Support/Claude/claude-code/claude --dangerously-skip-permissions
-102 101 Mon Jul  6 09:00:01 2026 /Users/test/Library/Application Support/Claude/claude-code/claude --dangerously-skip-permissions
-201   1 Mon Jul  6 09:01:00 2026 /opt/homebrew/bin/codex exec --full-auto strange argv here
-202   1 Mon Jul  6 09:02:00 2026 /Applications/Codex.app/Contents/Resources/codex app-server --listen stdio
-203   1 Mon Jul  6 09:03:00 2026 /usr/local/bin/codex --help
-301   1 Mon Jul  6 09:04:00 2026 /Users/test/.local/bin/claude-code-acp --stdio
-401   1 Mon Jul  6 09:05:00 2026 /Applications/Codex.app/Contents/Frameworks/Codex Framework.framework/Helpers/Codex (Renderer) --type=renderer
-bad line
-";
-
-        let records = AgentPSOutputParser::parse(output);
-        let agents = AgentPSOutputParser::agent_processes(&records);
-
-        assert_eq!(agents.len(), 2);
-    }
-
-    #[test]
-    fn session_scan_config_cuts_off_active_and_file_only_windows() {
-        let config = SessionScanConfig::default();
-        let now = Utc.with_ymd_and_hms(2026, 7, 12, 0, 0, 0).unwrap();
-
-        assert_eq!(
-            config.state(Some(now - chrono::Duration::seconds(119)), now, true),
-            AgentSessionState::Active
-        );
-        assert_eq!(
-            config.state(Some(now - chrono::Duration::seconds(121)), now, true),
-            AgentSessionState::Idle
-        );
-        assert!(config.file_only_session_allowed(now - chrono::Duration::minutes(29), now));
-        assert!(!config.file_only_session_allowed(now - chrono::Duration::minutes(31), now));
-    }
-
-    #[test]
-    fn agent_session_round_trips_json() {
-        let session = AgentSession {
-            id: "session-1".to_string(),
-            provider: AgentSessionProvider::Codex,
-            source: AgentSessionSource::DesktopApp,
-            state: AgentSessionState::Active,
-            pid: Some(1234),
-            transcript_path: Some("C:\\sessions\\rollout.jsonl".to_string()),
-            host: "devbox".to_string(),
-            workspace: AgentSessionWorkspace {
-                cwd: Some("C:\\work\\proj".to_string()),
-                project_name: Some("proj".to_string()),
-            },
-            activity: AgentSessionActivity {
-                started_at: Some(Utc.with_ymd_and_hms(2026, 7, 12, 0, 0, 0).unwrap()),
-                last_activity_at: Some(Utc.with_ymd_and_hms(2026, 7, 12, 0, 1, 0).unwrap()),
-            },
-            focus_target: AgentSessionFocusTarget::Process { pid: 1234 },
-        };
-
-        let json = serde_json::to_string(&session).unwrap();
-        let round_tripped: AgentSession = serde_json::from_str(&json).unwrap();
-        assert_eq!(round_tripped, session);
-        assert!(json.contains("\"focusTarget\""));
-    }
-
-    #[test]
-    fn focus_result_serializes_safely() {
-        let focused = serde_json::to_value(&SessionFocusResult::Focused).unwrap();
-        let unsupported = serde_json::to_value(&SessionFocusResult::Unsupported {
-            message: "focus unavailable".to_string(),
-        })
-        .unwrap();
-        let failed = serde_json::to_value(&SessionFocusResult::Failed {
-            message: "failed to focus".to_string(),
-        })
-        .unwrap();
-
-        assert!(focused.is_string() || focused.is_object());
-        assert_eq!(unsupported["message"], "focus unavailable");
-        assert_eq!(failed["message"], "failed to focus");
-    }
-
-    #[test]
-    fn host_validation_dedupes_and_rejects_unsafe_values() {
-        let hosts = RemoteSessionFetcher::sanitized_hosts(&[
-            "".to_string(),
-            " ".to_string(),
-            "-bad".to_string(),
-            "good".to_string(),
-            "good".to_string(),
-            "GOOD".to_string(),
-            "bad host".to_string(),
-            "bad\tcontrol".to_string(),
-        ]);
-        assert_eq!(hosts, vec!["good".to_string()]);
-    }
-
-    #[test]
-    fn codex_rollout_parser_reads_first_line_metadata() {
-        let metadata = CodexRolloutFirstLineParser::parse(
-            r#"{"type":"session_meta","payload":{"session_id":"abc","cwd":"C:\\work\\proj","originator":"codex_exec","source":"cli"}}"#,
-        )
-        .unwrap();
-        assert_eq!(metadata.session_id, "abc");
-        assert_eq!(metadata.cwd.as_deref(), Some("C:\\work\\proj"));
-    }
-
-    #[test]
-    fn claude_cwd_escape_is_stable() {
-        assert_eq!(
-            ClaudeSessionProjectMapper::escaped_cwd(r"C:\Users\me\My Project!"),
-            "C--Users-me-My-Project-"
-        );
-    }
-
-    #[test]
-    fn remote_session_result_round_trips() {
-        let result = AgentSessionHostResult {
-            host: "devbox".to_string(),
-            sessions: vec![AgentSession {
-                id: "session-1".to_string(),
-                provider: AgentSessionProvider::Claude,
-                source: AgentSessionSource::Cli,
-                state: AgentSessionState::Idle,
-                pid: None,
-                transcript_path: None,
-                host: "devbox".to_string(),
-                workspace: AgentSessionWorkspace {
-                    cwd: None,
-                    project_name: None,
-                },
-                activity: AgentSessionActivity {
-                    started_at: None,
-                    last_activity_at: None,
-                },
-                focus_target: AgentSessionFocusTarget::None,
-            }],
-            error: Some("ssh not found".to_string()),
-        };
-
-        let json = serde_json::to_string(&result).unwrap();
-        let decoded: AgentSessionHostResult = serde_json::from_str(&json).unwrap();
-        assert_eq!(decoded, result);
-    }
-
-    #[test]
-    fn windows_process_parser_uses_metadata_and_ignores_command_lines() {
-        let output = r#"[
-            {
-                "ProcessId": 101,
-                "ParentProcessId": 1,
-                "CreationDate": "/Date(1783773458193)/",
-                "Name": "claude.exe",
-                "ExecutablePath": "C:\\Tools\\claude.exe",
-                "CommandLine": "claude.exe --api-key super-secret"
-            },
-            {
-                "ProcessId": 202,
-                "ParentProcessId": 1,
-                "CreationDate": "2026-07-12T00:02:03Z",
-                "Name": "codex.exe",
-                "ExecutablePath": "C:\\Tools\\codex.exe"
-            }
-        ]"#;
-
-        let records = WindowsProcessOutputParser::parse(output);
-
-        assert_eq!(records.len(), 2);
-        assert_eq!(records[0].provider, Some(AgentSessionProvider::Claude));
-        assert_eq!(records[1].provider, Some(AgentSessionProvider::Codex));
-        assert!(records[0].started_at.is_some());
-        assert_eq!(records[0].executable, "claude.exe");
-        assert!(!records[0].executable.contains("super-secret"));
-    }
-
-    #[test]
-    fn windows_process_query_never_requests_raw_command_lines() {
-        let options = LocalAgentSessionScanner::process_options(Duration::from_secs(2));
-        let script = options.extra_args.last().unwrap();
-
-        assert!(script.contains("ProcessId"));
-        assert!(script.contains("ExecutablePath"));
-        assert!(!script.contains("CommandLine"));
-    }
-
-    #[test]
-    fn codex_discovery_only_builds_today_and_yesterday_directories() {
-        let now = Utc.with_ymd_and_hms(2026, 7, 12, 1, 2, 3).unwrap();
-        let root = Path::new(r"C:\Users\me\.codex\sessions");
-
-        let directories = LocalAgentSessionScanner::codex_day_directories(root, now.date_naive());
-
-        assert_eq!(
-            directories,
-            vec![
-                root.join("2026").join("07").join("12"),
-                root.join("2026").join("07").join("11"),
-            ]
-        );
-    }
-
-    #[test]
-    fn claude_metadata_parser_stops_after_session_and_cwd_are_known() {
-        struct FirstLineOnly {
-            bytes: &'static [u8],
-            offset: usize,
-        }
-
-        impl Read for FirstLineOnly {
-            fn read(&mut self, buffer: &mut [u8]) -> io::Result<usize> {
-                assert!(
-                    self.offset < self.bytes.len(),
-                    "parser read past complete metadata"
-                );
-                buffer[0] = self.bytes[self.offset];
-                self.offset += 1;
-                Ok(1)
-            }
-        }
-
-        let input = FirstLineOnly {
-            bytes:
-                b"{\"type\":\"user\",\"sessionId\":\"session-1\",\"cwd\":\"C:\\\\work\\\\proj\"}\n",
-            offset: 0,
-        };
-
-        let metadata = ClaudeTranscriptMetadataParser::parse(input).unwrap();
-
-        assert_eq!(metadata.session_id.as_deref(), Some("session-1"));
-        assert_eq!(metadata.cwd.as_deref(), Some(r"C:\work\proj"));
-    }
-
-    #[test]
-    fn focus_rejects_remote_and_file_only_targets_explicitly() {
-        let remote = AgentSession {
-            id: "remote".to_string(),
-            provider: AgentSessionProvider::Codex,
-            source: AgentSessionSource::Cli,
-            state: AgentSessionState::Active,
-            pid: Some(42),
-            transcript_path: None,
-            host: "other-host".to_string(),
-            workspace: AgentSessionWorkspace {
-                cwd: None,
-                project_name: None,
-            },
-            activity: AgentSessionActivity {
-                started_at: None,
-                last_activity_at: None,
-            },
-            focus_target: AgentSessionFocusTarget::Process { pid: 42 },
-        };
-        let file_only = AgentSession {
-            host: "localhost".to_string(),
-            focus_target: AgentSessionFocusTarget::Transcript {
-                transcript_path: "session.jsonl".to_string(),
-            },
-            ..remote.clone()
-        };
-
-        assert!(matches!(
-            focus_session(&remote),
-            SessionFocusResult::Unsupported { .. }
-        ));
-        assert!(matches!(
-            focus_session(&file_only),
-            SessionFocusResult::Unsupported { .. }
-        ));
-    }
-
-    #[test]
-    fn ssh_fetch_uses_noninteractive_options_and_a_strict_timeout() {
-        let options =
-            RemoteSessionFetcher::ssh_options("user@devbox", Duration::from_secs(5)).unwrap();
-
-        assert_eq!(options.timeout, Duration::from_secs(5));
-        assert!(
-            options
-                .extra_args
-                .windows(2)
-                .any(|pair| pair == ["-o", "BatchMode=yes"])
-        );
-        assert!(
-            options
-                .extra_args
-                .windows(2)
-                .any(|pair| pair == ["-o", "ConnectTimeout=3"])
-        );
-        assert!(options.extra_args.iter().any(|arg| arg == "user@devbox"));
-    }
-
-    #[tokio::test]
-    async fn ssh_hosts_are_fetched_in_parallel_and_failures_are_isolated() {
-        let barrier = Arc::new(tokio::sync::Barrier::new(2));
-        let results = tokio::time::timeout(
-            Duration::from_secs(1),
-            RemoteSessionFetcher::fetch_hosts_with(
-                &[
-                    "slow-a".to_string(),
-                    "slow-b".to_string(),
-                    "broken".to_string(),
-                ],
-                |host| {
-                    let barrier = Arc::clone(&barrier);
-                    async move {
-                        if host == "broken" {
-                            return AgentSessionHostResult::failed(host, "unreachable");
-                        }
-                        barrier.wait().await;
-                        AgentSessionHostResult::success(host, Vec::new())
-                    }
-                },
-            ),
-        )
-        .await
-        .expect("hosts were fetched sequentially");
-
-        assert_eq!(results.len(), 3);
-        assert_eq!(
-            results
-                .iter()
-                .filter(|result| result.error.is_some())
-                .count(),
-            1
-        );
-        assert_eq!(
-            results
-                .iter()
-                .filter(|result| result.error.is_none())
-                .count(),
-            2
-        );
-    }
-
-    #[tokio::test]
-    async fn disabled_discovery_returns_without_running_scanners() {
-        let result = AgentSessionDiscovery::default()
-            .scan(AgentSessionDiscoveryMode::Disabled)
-            .await;
-
-        assert_eq!(result, AgentSessionDiscoveryResult::Disabled);
-    }
-}
+include!("agent_sessions/tests.rs");
